Include default processing times config in scan config route
Configuration of allowed processing times is required by the
delivery dashboard to show effects of rescorings on due dates.
However, this configuration is optional and does not have to
be included in the scan configuration. Nevertheless, if it is
not included, there is a fallback to the static default config
anyways, so delivery dashboard should use it as well.