Add manual-release pipeline bypassing CI
As CI (in that: test and linter-suite) is running on latest
delivery-service image, we have a conceptional problem if there are
required changes on upstream dependencies (e.g. cc-utils), the CI steps
will fail.
In the past we solved those by temporally disabling CI steps on release
pipeline.

This change at least removes the necessity to modifying pipeline(s) to
solve these edge-cases.