Don't require artefact nodes to exist for issue replicator
Until now, we could assume each artefact in a backlog item refers to one
artefact which is referenced in a component descriptor, i.e. a resource
or a source. However, the recently added kind "runtime" does not fit
into this schema, since it is not included in the OCM. So, we have to
widen the allowed artefacts and must not assume they can be found in a
component descriptor and thus we cannot use `ArtefactNodes` as common
data model. Instead, always try to use `ComponentArtefactId` which
already support artefacts of kind "runtime" as well. In those cases,
where extra information from the component descriptor is required, e.g.
labels or accesses, add those by retrieving the component descriptor
ad-hoc.