Display artefact versions without scan in GitHub issue
This also fixes a bug where the "scan-pending" semantic did not work
properly in case multiple versions of the same artefact are in the
scanned BoM, while some of the versions were scanned and others were
not. With this change, an issue gets the "scan-pending" label if at
least one of the contained artefact versions is not scanned yet (thus
keeping the issue open although no finding might be found yet). Before,
only issues were all contained artefact versions had no scan were
labeled and considered as "scan-pending".