Fallback to artefact query if component version is not set
If the component version of a database entry is not specified and `none_ok` is not `True`,
it is checked whether the component in question contains an artefact version which matches
the database entry. This is especially useful for retrieving BDBA scan results which don't
contain a component version (for deduplication), to only query scan results for artefact
versions which are included in the specified component versions.