Overwrite issue milestone if issue is re-opened
Normally, an issue can't be assigned to a milestone which is already
closed. However, if an closed issue is assigned to a milestone which is
also closed, and the issue is re-opened, the milestone persists.
In the reporting, the milestone is normally not modified once it is set
to allow manual re-assignment to another milestone. However, in case the
issue was closed before, it can be assumed that the manual re-assignment
is not of interest anymore, but rather the assignment to a milestone
which is still open.