Add cors middleware as first middleware
Middlewares are being executed in the following order:
- first (pre-request-handling)
- second (pre-request-handling)
- request-handling
- second (post-request-handling)
- first (post-request-handling)

Previously, the auth middlware was added "first" and the cors middleware
"second". In this scenario, the cors middleware was never considered if
the auth middleware resulted in an error case during
pre-request-handling (e.g. if the request was unauthenticated). Because
of that, auth-related errors did not contain the expected cors headers.