Also match DB query if component version is not specified
To allow querying all entries with a given component name but
independent of the component version, match any component version if it
is not specified by the caller. This is required for the issue
replicator which queries for all compliance snapshots of a given
artefact, independent of its concrete version (it is aggregated across
multiple versions).