Invalidate compliance summary cache upon new artefact scan info
Just invalidating the compliance summary with the same datatype as the
metadata entry which is currently being updated is not working as
desired in case the entry in question is of type
`meta/artefact_scan_info`. In that case, the compliance summary for all
finding types related to the datasource have to be updated, i.e. to
update the summary from UNKNOWN (no scan exists) to CLEAN/FINDINGS.