Don't keep package hint function unnecessarily overflown
Already since the refactoring of BDBA scans as OCM-Gear extension, there
will always be just one scan result at a time because of the processing
via BLIs. Hence, there is no need to check if the result matches the
given resource (it was just previously retrieved exactly for the same
resource). Also, as only resources are being scanned, there is no need
to allow passing-in of arbitrary artefacts.