Don't propose any severities if finding is already rescored
Before, the severity of the rescoring proposal was either the original
severity or the automatically resored one (based on the configured
rescoring rules). However, there is no need to try to propose any
severity in case the finding is already rescored. In that case, just use
the current (in that: rescored) severity. The reference to the original
severity is still stored under the `finding` property.