Consider all sprints for creation of GitHub milestones
If the processing date of a finding lies too much in the past and we
don't consider enough sprints, it may happen that an issue is not
assigned to any sprint because no related milestone is kept open. By
considering all available sprints, it will always be assigned to a
sprint (assuming there is at least one not-closed milestone).