Be more defensive in autoscaling policies
Actual workload has shown that it is more useful to scale steady instead
of too much at once. This is because the Kubernetes metrics (i.e. CPU
consumption) are provided with a little delay (default is 60s), so in
case there is very temporary load peak which is visible in the metrics,
scaling-up e.g. 100% of pods is not beneficial if this load was already
handled as soon as the updated metric becomes available. Hence, rather
scale steady to be able to properly adapt to more general load peaks and
prevent unnecessary extreme scaling-behaviours.