Only use sprints from compliance snapshots for issue replication
Compliance snapshots are only being created for sprints of a configured
date interval. However, as of before this change, the issue replicator
considered all sprints returned by the delivery-service api and grouped
the open findings among these sprints. As not all of these sprints must
have a respective compliance snapshots (i.e. if the sprint is not part
of the configured date interval), the finding is not being reported in
the GitHub issues.
To fix this, this change only considers the sprints determined from the
compliance snapshots and groups the findings among these sprints. That
way, the configured interval for the artefact enumerator is the single
sourth of truth and all findings are being reported only for these
sprints.