Use `aws_cfg_name` instead of `aws_cfg_set_name` for BDBA
As secrets are no longer managed using configuration sets, directly pass
the name of the aws secret instead of only referencing the umbrella
configuration set and requiring it to have a default aws config. Note:
This also ensures consistency towards the clamav extension as it also
requires the `aws_cfg_name` (!= `aws_cfg_set_name`) to be specified.