Add `SecretFactory` and respective model classes
The `SecretFactory` is similar to Gardener-CICD's `ConfigFactory`,
especially with regards to its consumption, however, it drops
assumptions and requirements which are only valid for Gardener-CICD and
not for the OCM-Gear as such, which allows dropping dependencies and
complexity. Also, it focuses on the consumption of secrets via
Kubernetes secrets (mounted into the filesystem at `SECRET_FACTORY_PATH`).

To ensure backwards compatibility (for now), the serialised
`ConfigFactory` at `CFG_FACTORY_SECRET_PATH` is used as fallback to
construct the `SecretFactory` from it. This allows the deployment to
stay as-is until the Kubernetes secrets are being properly created.

Also, overwriting `secret_type_to_class` upon initialisation of the
`SecretFactory` allows OCM-Gear extensions to add custom secret types
and model classes without the need to track them centrally within this
`secret_mgmt` package.