Add dedicated secrets as volume mounts
As of now, it is still possible and allowed to use the
`cfg-factory-secret` to store the required credentials and mount it at
`CFG_FACTORY_SECRET_PATH`. Because of that, these newly added mounts
have to be optional at first, until support for the `cfg-factory-secret`
has been removed (marked with "TODO" in this commit).