Stop re-applying existing BDBA triages to new BDBA apps
As of now, BDBA triages are stored as rescorings in the delivery-db
using the `SINGLE` scope, i.e. they are not re-used across new artefact
versions. With this change, the scope is extended to `ARTEFACT`, so they
will be also applied to new versions, however, not among different
artefact extra ids (this is different to now!).

With that, we no longer have to also re-apply existing BDBA triages to
new BDBA apps, as we use the rescorings in the delivery-db as single
sourth of truth anyways (the BDBA UI may show findings already which
actually don't have to be assessed anymore because they were already
assessed using the delivery-dashboard).

This also elimnates the requirement to delete old BDBA apps, as it was
required as an optimisation to have less apps from where to import
triages. Since we don't retrieve every existing app result anymore, this
isn't a limitation anymore.

With this change, the BDBA config properties `reference_group_ids` and
`delete_inactive_products_after_seconds` are not required anymore and
thus removed from the scan configuration.