Add refactored scan cfg and new finding-specific model classes
As a first step, these model classes are kept totally redundant to the
current ones defined in `config.py` to allow a step-by-step transition
towards the overhauled cfg-mgmt. This will allow existing extensions and
in the end the delivery-service + -dashboard themselves to switch to the
overhauled cfg-mgmt one by one as long as the provided configuration is
kept semantically equal.