Use mounted scan cfg for cache manager
With this change, the cache manager will retrieve the scan cfg as well
as the finding cfgs from a local or mounted file instead of the custom
resource. Therefore, it requires either setting of the arguments
`--scan-cfg-path` and `findings-cfg-path` or the env vars `SCAN_CFG_PATH`
and `FINDINGS_CFG_PATH` instead of `--cfg-name` or `CFG_NAME`.