Make `cfg_name` optional for backlog item claim
Since we are in the process of allowing multiple scan cfgs to requiring
a singleton, the `cfg_name` will be dropped one after another for each
extension. Hence, allow the artefact enumerator to keep creating backlog
items with set cfg names (for now), but also allow already migrated
extensions to ignore the cfg name and process all BLIs independent of
it.