Remove obsolete `auto_triage` for BDBA scans
The option for auto-triaging was added for use cases where components
should be only scanned for licenses and not for vulnerabilities, thus
all vulnerability findings were just auto-triaged. Now, with the
finding-cfgs, it is possible to configure certain components only for
licenses and not for vulnerabilities, hence this auto-triaging is not
required anymore.

Also, as the auto-triaging was indicated using the
`cloud.gardener.cnudie/dso/scanning-hints/binary_id/v1` label, this
label can be removed eventually.