Fix: Close issues which "fell out" of the configured time range
The artefact enumerator is configured with a time range for which it
creates respective compliance snapshots. As these compliance snapshots
(or more precisely: their correlation id), is used as a reference to the
GitHub issues, only issues which have a correlation id for which a
compliance snapshot exists are being considered for the automatic issue
lifecycle. Now, if a compliance snapshot becomes inactive and no new
snapshot is created with the same correlation id (because by now the due
date is not part of the configured interval anymore), the respective
GitHub issue is not considered anymore by the automatic issue lifecycle.

To mitigate this, this fix will consider all sprints (for active and
inactive snapshots), but will only assign findings to the active ones,
whereby GitHub issues for the inactive ones will be closed because they
have no assigned findings anymore.