Disable SQLAlchemy DB session pool size limit
The PostgreSQL helm chart allows configuration of the maximum connections
via the helm value `postgresqlMaxConnections` (which can be set by an
operator installing an instance of the ODG). However, SQLAlchemy
simultaneously defines an own connection limit for a session object of 5
(+ an overflow of 10).
As we see an increasing number of extensions, interacting simultaneously
with the Delivery-Service(-DB)-API, the default db session pool size and
the allowed overflow are not enough anymore, in case many requests are
routed to the same instance of the Delivery-Service. This may ultimately
result in MaxConnectionErrors being raised. As a mitigation, disable
limiting on SQLAlchemy side and rely on PostgreSQL's configuration of
`postgresqlMaxConnections`.