Raise `503 service unavailable` error upon DB connection timeout
A DB connection timeout error is raised in case the session pool is
full, i.e. the pool size of "5" and the possible overflow of "10" is
exceeded. In that case, raise a http 503 error so that the load balancer
(ingress nginx controller) is able to forward the request to the next
available pod which might have a session pool capacity left.