Remove `automatic-release` job
The delivery-service has less and less dependencies towards cc-utils by
now (and this trend is expected to continue in the future). Also, in
case changes in cc-utils are required, there are most commonly also
changes required in the delivery-service as well, hence there is a
manual intervention anyways. Because of that, stop auto-releasing upon
dependency updates.