Add `access-manager` extension
The access manager is intended to be run as a regular cronjob which takes care
of updating the user role bindings in the database. This is required to reflect
changes in the origin of the respective role bindings, e.g. a change of the
GitHub organisation and/or team membership.