Add required odg-extensions from regular extension-cfg
odg-operator is designed as regular odg extension.
Thus, it has access to common extension cfg.

To offer a great out-of-box experience, an enduser should not be
required to bother with odg internals. Thus, the odg-extensions which
are part of the required bare minimum for an odg to be an odg, should be
configured somewhere "else".
This "else" is the extension cfg, which is used to derive the final set
of to-be deployed odg extensions.

Signed-off-by: Philipp Heil (zkdev) <philipp.heil@sap.com>