feat(modg): Deploy delivery-db password as secret
postgres requires a password which must either be templated into the
deployment itself, or be referenced via a secret.
first variant is no option, because the manifest is templated centrally
(odg-operator) and has no user secrets at hand. thus, deploy a secret in
user cluster (as part of bootstrap chart) which the postgres deployment
can use to extract the password.

Signed-off-by: Philipp Heil (zkdev) <philipp.heil@sap.com>