Deduplicate rescorings in issue-replicator
In case multiple artefacts (which are being processed for the same
backlog item) contain the very same rescoring (e.g. because of a more
unspecific rescoring scope), this rescoring will be retrieved multiple
times (i.e. for each artefact). Because the "historical findings"
displayed in the GitHub issues are calculated using the rescorings, this
causes the rescoring to be considered multiple times for this history.
Instead, deduplicate rescorings to ensure they are only considered once.

Signed-off-by: Jonas Brand (8R0WNI3) <j.brand@sap.com>